Embedding actors only via the options menu
With this upcoming release, actors can only be embedded into the scene via the options menu.  A list of unique actors will be provided to choose from, or all actors in the list can be embedded.  The actor in the actor folder will not be deleted.

If the scene is used while the actor is still in the actors folder, token mold ignores to token and core functionality is maintained.  Token Independence will only affect tokens that have "null" actors, ie no actor in the actors folder.

The old method of embedding by deleting the actor has been removed from the code.
This module will no longer clean-up the scene when the last token for a given actor is deleted, since there is now an option in the menu to list and delete selected embedded actors.